Initial Check-In
Multi-threaded proxy application. Will create two IO connections (To the miner and to the proxy) on separate threads. 

The application will read from the miner and pass on the information to the pool and vice versa. 

System supports plain text Stratum V1 and V2 protocols. SSL is not supported at this time.