Error Code Response
When we respond to the miner with an error code the pool no longer puts items in the stream. So, we never write to the miner. Therefore we never know that the miner has disconnected and we will wont' dispose of the items. A check if the connection is still alive via trying to write a single byte of data to the stream will let us know if it needs to be terminated.