Statistics Feature
Logging all JsonRpc messages is no longer the primary way of displaying. That only works when debug is on now. We now have a display that shows connected miners and their associated info. 

Tested with multiple threads
Tested with multiple connections and losses
Tested with Stratum version 1 & 2

Still does not support SSL